% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.r
\name{bayes}
\alias{bayes}
\title{Bayes model}
\usage{
bayes(y, X, model = c("BayesRR", "BayesA", "BayesLASSO", "BayesB",
  "BayesBpi", "BayesC", "BayesCpi", "BayesR"), map = NULL, pi = NULL,
  fold = NULL, niter = 50000, nburn = 25000, windsize = NULL,
  wppa = 0.01, vara = NULL, dfvara = NULL, s2vara = NULL,
  vare = NULL, dfvare = NULL, s2vare = NULL, outfreq = 100,
  seed = 666666, verbose = TRUE)
}
\arguments{
\item{y}{vector of phenotype, NAs are not allowed.}

\item{X}{numeric matrix of genotype with individuals in rows and markers in columns, NAs are not allowed.}

\item{model}{bayes model including: "BayesRR", "BayesA", "BayesLASSO", "BayesB", "BayesBpi", "BayesC", "BayesCpi", "BayesR".}

\item{map}{(optional, only for GWAS) the map information of genotype, columns are: SNPs, chromosome, physical position.}

\item{pi}{percentage of zero effect SNPs. For bayesR, it is a vector for groups of SNPs, the default is c(0.95, 0.02, 0.02, 0.01).}

\item{fold}{percentage of variance explained for groups of SNPs, the default is c(0, 0.0001, 0.001, 0.01).}

\item{niter}{the number of MCMC iteration.}

\item{nburn}{the number of iterations to be discarded.}

\item{windsize}{window size in bp for GWAS, the default is 1e6.}

\item{wppa}{the threshold of genetic variance explained by single window, the default is 0.01.}

\item{vara}{prior value of genetic variance.}

\item{dfvara}{the number of degrees of freedom for the distribution of genetic variance.}

\item{s2vara}{scale parameter for the distribution of genetic variance.}

\item{vare}{prior value of residual variance.}

\item{dfvare}{the number of degrees of freedom for the distribution of residual variance.}

\item{s2vare}{scale parameter for the distribution of residual variance.}

\item{outfreq}{frequency of information output on console, the default is 100.}

\item{seed}{seed for random sample.}

\item{verbose}{whether to print the iteration information.}
}
\description{
Bayes linear regression model using individual level data
}
\examples{
#' need to update
}
