% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plink.r
\name{read_plink}
\alias{read_plink}
\title{data load}
\usage{
read_plink(
  bfile = "",
  maxLine = 10000,
  impute = TRUE,
  mode = c("A", "D"),
  out = NULL,
  threads = 0
)
}
\arguments{
\item{bfile}{character, prefix of Plink binary format data.}

\item{maxLine}{number, set the number of lines to read at a time.}

\item{impute}{logical, whether to impute missing values in genotype by major alleles.}

\item{mode}{"A" or "D", additive effect or dominant effect.}

\item{out}{character, path and prefix of output file}

\item{threads}{number, the number of used threads for parallel process}
}
\description{
To load plink binary data
}
\examples{
bfile_path = system.file("extdata", "example", package = "hibayes")
data = read_plink(bfile_path, out=tempfile(), mode="A")
pheno = data$pheno
geno = data$geno
map = data$map
}
