% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{CMplot}
\alias{new}
\alias{packageVersion}
\alias{read.delim}
\alias{read.table}
\alias{write.table}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{CMplot}{\code{\link[CMplot:CMplot-package]{CMplot}}}

  \item{methods}{\code{\link[methods]{new}}}

  \item{utils}{\code{\link[utils:packageDescription]{packageVersion}}, \code{\link[utils:read.table]{read.delim}}, \code{\link[utils]{read.table}}, \code{\link[utils]{write.table}}}
}}

