% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssbayes.r
\name{ssbayes}
\alias{ssbayes}
\title{Single-step Bayes model}
\usage{
ssbayes(
  y,
  X,
  Ped,
  mode = c("A", "D"),
  model = c("BayesB", "BayesA", "BayesLASSO", "BayesRR", "BayesBpi", "BayesC",
    "BayesCpi", "BayesR"),
  map = NULL,
  pi = NULL,
  fold = NULL,
  niter = 20000,
  nburn = 12000,
  windsize = NULL,
  wppa = 0.01,
  vg = NULL,
  dfvg = NULL,
  s2vg = NULL,
  ve = NULL,
  dfve = NULL,
  s2ve = NULL,
  outfreq = 10,
  seed = 666666,
  verbose = TRUE
)
}
\arguments{
\item{y}{vector of phenotype, NAs are not allowed.}

\item{X}{numeric matrix of genotype with individuals in rows and markers in columns, NAs are not allowed.}

\item{model}{bayes model including: "BayesRR", "BayesA", "BayesLASSO", "BayesB", "BayesBpi", "BayesC", "BayesCpi", "BayesR".}

\item{map}{(optional, only for GWAS) the map information of genotype, columns are: SNPs, chromosome, physical position.}

\item{pi}{percentage of zero effect SNPs. For bayesR, it is a vector for groups of SNPs, the default is c(0.95, 0.02, 0.02, 0.01).}

\item{fold}{percentage of variance explained for groups of SNPs, the default is c(0, 0.0001, 0.001, 0.01).}

\item{niter}{the number of MCMC iteration.}

\item{nburn}{the number of iterations to be discarded.}

\item{windsize}{window size in bp for GWAS, the default is NULL.}

\item{wppa}{the threshold of genetic variance explained by single window, the default is 0.01.}

\item{vg}{prior value of genetic variance.}

\item{dfvg}{the number of degrees of freedom for the distribution of genetic variance.}

\item{s2vg}{scale parameter for the distribution of genetic variance.}

\item{ve}{prior value of residual variance.}

\item{dfve}{the number of degrees of freedom for the distribution of residual variance.}

\item{s2ve}{scale parameter for the distribution of residual variance.}

\item{outfreq}{frequency of information output on console, the default is 100.}

\item{seed}{seed for random sample.}

\item{verbose}{whether to print the iteration information.}
}
\description{
Single-step Bayes linear regression model using individual level data and pedigree information
}
\examples{
set.seed(123)
n = 1000                        # number of observations
m = 100                         # number of SNPs
k = 10                          # number of causal SNPs
h2 = 0.5                        # heritability
X = matrix(sample(c(0, 1, 2), n*m, prob=c(0.25, 0.5, 0.25), replace=TRUE), n)
X = apply(X, 2, function(x){x-mean(x)})
qtl = sort(sample(1:m, k))
betaTrue = array(0,m)
betaTrue[qtl] = rnorm(k)
g = X\%*\%betaTrue
vg = var(g)
ve = (1-h2)/h2 * vg
y = g + rnorm(n,0,sqrt(ve))

## fit model
fit = ssbayes(y=y, X=X, model="BayesCpi", niter=1000, nburn=500)

## check results
cor(fit$g, betaTrue)
plot(fit$g, betaTrue)

#####load plink binary file#####
# bfile_path = system.file("extdata", "example", package = "hibayes")
# data = read_plink(bfile_path, out=tempfile())
# pheno = data$pheno
# geno = data$geno
# map = data$map
## For GS/GP
# fit = ssbayes(y=pheno[,1], X=geno, model="BayesR", niter=20000, nburn=10000)
## For GWAS
# fit = ssbayes(y=pheno[,1], X=geno, map=map, windsize=1e6, model="BayesR", niter=20000, nburn=10000)
}
