% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plink.r
\name{read_plink}
\alias{read_plink}
\title{data load}
\usage{
read_plink(bfile = "", maxLine = 10000, mode = c("A", "D"),
  backingpath = NULL, descriptorfile = NULL, backingfile = NULL,
  threads = 1)
}
\arguments{
\item{bfile}{character, prefix of Plink binary format data.}

\item{maxLine}{number, set the number of lines to read at a time.}

\item{mode}{"A" or "D", additive effect or dominant effect.}

\item{backingpath}{the path to the directory containing the file backing cache.}

\item{descriptorfile}{the name of the file to hold the backingfile description, for subsequent use with ‘attach.big.matrix’; if ‘NULL’, the ‘backingfile’ is used as the root part of the descriptor file name.  The descriptor file is placed in the same directory as the backing files.}

\item{backingfile}{the root name for the file(s) for the cache.}

\item{threads}{number, the number of used threads for parallel process}
}
\description{
To load plink binary data
}
\examples{
bfile_path = system.file("extdata", "example", package = "hibayes")
data = read_plink(bfile_path, mode="A")
pheno = data$pheno
geno = data$geno
map = data$map
}
