% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plink.r
\name{read_plink}
\alias{read_plink}
\title{data load}
\usage{
read_plink(bfile = "", maxLine = 10000, mode = c("A", "D"),
  threads = 1)
}
\arguments{
\item{bfile}{character, prefix of Plink binary format data.}

\item{maxLine}{number, set the number of lines to read at a time.}

\item{mode}{"A" or "D", additive effect or dominant effect.}

\item{threads}{number, the number of used threads for parallel process}
}
\description{
To load plink binary data
}
\examples{
need to update
}
