% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbayes.r
\name{sbayes}
\alias{sbayes}
\title{SBayes model}
\usage{
sbayes(
  sumstat,
  ldm,
  model = c("SBayesB", "SBayesA", "SBayesLASSO", "SBayesRR", "SBayesBpi", "SBayesC",
    "SBayesCpi", "SBayesR", "CG"),
  map = NULL,
  pi = NULL,
  lambda = NULL,
  fold = NULL,
  niter = 20000,
  nburn = 12000,
  windsize = NULL,
  wppa = 0.01,
  vara = NULL,
  dfvara = NULL,
  s2vara = NULL,
  vare = NULL,
  dfvare = NULL,
  s2vare = NULL,
  outfreq = 10,
  seed = 666666,
  verbose = TRUE
)
}
\arguments{
\item{sumstat}{matrix of summary data, details refer to https://cnsgenomics.com/software/gcta/#COJO.}

\item{ldm}{dense or sparse matrix, ld for reference panel (m * m, m is the number of SNPs), note that the order of SNPs should be consistent with summary data.}

\item{model}{sbayes model including: "SBayesRR", "SBayesA", "SBayesLASSO", "SBayesB", "SBayesBpi", "SBayesC", "SBayesCpi", "SBayesR", "CG".}

\item{map}{(optional, only for GWAS) the map information of genotype, columns are: SNPs, chromosome, physical position.}

\item{pi}{percentage of zero effect SNPs. For bayesR, it is a vector for groups of SNPs, the default is c(0.95, 0.02, 0.02, 0.01).}

\item{lambda}{value or vector, the ridge regression value for each SNPs.}

\item{fold}{percentage of variance explained for groups of SNPs, the default is c(0, 0.0001, 0.001, 0.01).}

\item{niter}{the number of MCMC iteration.}

\item{nburn}{the number of iterations to be discarded.}

\item{windsize}{window size in bp for GWAS, the default is 1e6.}

\item{wppa}{the threshold of genetic variance explained by single window, the default is 0.01.}

\item{vara}{prior value of genetic variance.}

\item{dfvara}{the number of degrees of freedom for the distribution of genetic variance.}

\item{s2vara}{scale parameter for the distribution of genetic variance.}

\item{vare}{prior value of residual variance.}

\item{dfvare}{the number of degrees of freedom for the distribution of residual variance.}

\item{s2vare}{scale parameter for the distribution of residual variance.}

\item{outfreq}{frequency of information output on console, the default is 100.}

\item{seed}{seed for random sample.}

\item{verbose}{whether to print the iteration information.}
}
\description{
Bayes linear regression model using summary level data
}
\examples{
bfile_path = system.file("extdata", "example", package = "hibayes")
data = read_plink(bfile_path, out=tempfile())
geno = data$geno
map = data$map
head(map)
sumstat_path = system.file("extdata", "example.ma", package = "hibayes")
sumstat = read.table(sumstat_path, header=TRUE)
head(sumstat)

## computate ld variance covariance matrix
# ldm1 = ldmat(geno, threads=4)   #chromosome wide full ld matrix
# ldm2 = ldmat(geno, map, ldchr=FALSE, chisq=5, threads=4)   #chromosome block + sparse ld matrix

## if the order of SNPs in genotype is not consistent with the order in sumstat file, 
## prior adjusting is necessary.
# indx = match(sumstat[, 1], map[, 1])
# ldm1 = ldm1[indx, indx]

## fit model
# fit = sbayes(sumstat=sumstat, ldm=ldm1, model="SBayesR")
}
